refactor(controlplane): publish usercontext/entities under pkg/
The identity context helpers (CurrentUser, CurrentAPIToken, CurrentOrg,
CurrentMembership) lived under app/controlplane/internal/, yet the public
app/controlplane/pkg/biz depends on them: AuditorUseCase.Dispatch resolves the
audit actor from those context values. Downstream modules that embed pkg/biz
therefore cannot populate the actor, and every audit entry that requires one is
dropped with "failed to get actor information".

Move the package to app/controlplane/pkg/usercontext/entities so it is
importable by those consumers. Import-path change only, no behaviour change.


Chainloop-Trace-Sessions: e6906bba-e9a1-468f-8d04-e3fe7ae231df
