+++ app/controlplane/internal/service/cascredential.go
@@ -19,8 +19,8 @@ import (
 	"context"
 
 	pb "github.com/chainloop-dev/chainloop/app/controlplane/api/controlplane/v1"
-	"github.com/chainloop-dev/chainloop/app/controlplane/internal/usercontext/entities"
 	"github.com/chainloop-dev/chainloop/app/controlplane/pkg/authz"
+	"github.com/chainloop-dev/chainloop/app/controlplane/pkg/usercontext/entities"
 	errors "github.com/go-kratos/kratos/v2/errors"
 	"github.com/google/uuid"
