+++ app/controlplane/internal/service/workflow.go
@@ -21,8 +21,8 @@ import (
 	"time"
 
 	v1 "github.com/chainloop-dev/chainloop/app/controlplane/api/jsonfilter/v1"
-	"github.com/chainloop-dev/chainloop/app/controlplane/internal/usercontext/entities"
 	"github.com/chainloop-dev/chainloop/app/controlplane/pkg/authz"
+	"github.com/chainloop-dev/chainloop/app/controlplane/pkg/usercontext/entities"
 	"github.com/chainloop-dev/chainloop/pkg/jsonfilter"
 
 	pb "github.com/chainloop-dev/chainloop/app/controlplane/api/controlplane/v1"
